Remove line break when no Reverse Dependencies
When there are no Reverse Dependencies to list, the word "none" was
being displayed on the next line. That could be confusing, because all
other single-line values are displayed on the same line as the header.
This removes the line break before "none" so it is displayed on the
same line as the header like other values.